Fix reconnect logic
Turns out publishReplay() also replays the 'complete' message and it
remembers it even after refCount() goes to zero. This means after a
complete it will never resubscribe the source observable.

Now use shareReplay with refcounting again and error() instead of
complete() when the websocket disconnects. Documentation states an error
will not be remembered, so it will resubscribe the source observable when
needed.